ml-models/requirements.txt
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.6.0
catboost==1.2.2
firebase-admin==6.2.0
python-dotenv==1.0.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
statsmodels==0.14.0
flask==3.0.0
gunicorn==21.2.0

